% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkem_main_functions.R
\name{F}
\alias{F}
\alias{dF}
\alias{rF}
\title{The (scaled) F Distribution}
\usage{
dF(x, df1, df2, beta, log = FALSE)

rF(n, df1, df2, beta)
}
\arguments{
\item{x}{vector of quantities.}

\item{df1}{First degrees of freedom}

\item{df2}{Second degrees of freedom}

\item{beta}{Scale parameter}

\item{log}{logical; if TRUE, density is given as log(p).}

\item{n}{number of draws}
}
\description{
Density and random generation for the F distribution with first degrees of freedom \code{df1},
second degrees of freedom \code{df2}, and scale parameter \code{beta}.
}
\examples{

draws_F <- rF(n=1e4, df1=2, df2=4, beta=1)
hist(draws_F,500,xlim=c(0,10),freq=FALSE)
seqx <- seq(0,10,length=1e5)
lines(seqx,dF(seqx, df1=2, df2=4, beta=1),col=2,lwd=2)

}
