% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkem_main_functions.R
\name{shrinkem}
\alias{shrinkem}
\title{Fast Bayesian regularization using Gaussian approximations}
\usage{
shrinkem(
  x,
  covmatrix,
  group = 1,
  type = "lasso",
  iterations = 10000,
  burnin = 10000,
  store = 10,
  sign.level = 0.95,
  df1 = 1,
  df2 = 1,
  scale2 = 1,
  ...
)
}
\description{
Fast Bayesian regularization using Gaussian approximations
}
